Change DTYPE to float64 & complex128

+++ QudiTop/global_var.py
@@ -3,7 +3,7 @@ Refer: https://gitee.com/forcekeng/quditop
 """
 import torch
 
-DTYPE = torch.float32
-CDTYPE = torch.complex64
+DTYPE = torch.float64
+CDTYPE = torch.complex128
 DEFAULT_VALUE = torch.tensor(0, dtype=DTYPE)
 DEFAULT_PARAM_NAME = '_param_'
